+++ docs/source/changelog.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+Version 0.2.0
+-------------
+
+Added the fritzCallMon.py script
+Made the fritzBackwardSearch script callable as library
+Renamed the script as dashes are not working in an import statement
+
 Version 0.1.0
 -------------